JungsiPredictionModel/pyproject.toml
[project]
name = "jungsi-prediction-model"
version = "1.0.0"
description = "정시 합격 예측 시스템 - ML + RAG 기반 실시간 입시 예측 플랫폼"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "JungsiPrediction", email = "contact@example.com"}
]

dependencies = [
    # Web Crawling
    "playwright>=1.40.0",
    "beautifulsoup4>=4.12.0",
    "httpx>=0.25.0",
    "fake-useragent>=1.4.0",

    # Database
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.0",

    # Google Cloud
    "google-cloud-aiplatform>=1.38.0",
    "google-cloud-storage>=2.14.0",
    "google-cloud-pubsub>=2.18.0",
    "google-cloud-documentai>=2.20.0",
    "google-generativeai>=0.3.0",

    # ML & Data
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.3.0",
    "xgboost>=2.0.0",
    "lightgbm>=4.1.0",
    "catboost>=1.2.0",
    "joblib>=1.3.0",

    # API
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "tenacity>=8.2.0",
    "schedule>=1.2.0",
    "redis>=5.0.0",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
